Update CSS comments
The margin of lists in IE6/7 is set by 'pt', not relative to the root
font-size. This is contrary to the HTML5 spec and other modern
browsers, including IE8+.

If people need to customize the list margin, they can either edit
normalize.css directly or override later in the source. But at least
any non-customised lists will behave consistently now.