Normalize typography
The margin of many elements in IE6/7 is set by 'pt', not relative to
the root font-size. This is contrary to the HTML5 spec and all other
modern browsers, including IE8+.

If people need to customize margins, they can either edit normalize.css
directly or override later in the source. But at least any
non-customized elements will behave consistently now.